requirements.txt
appnope==0.1.0
decorator==4.0.11
fire==0.1.0
h5py==2.7.0
ipython==5.3.0
ipython-genutils==0.2.0
Keras==2.0.2
numpy==1.12.1
olefile==0.44
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.0.0
prompt-toolkit==1.0.14
protobuf==3.2.0
ptyprocess==0.5.1
Pygments==2.2.0
PyYAML==3.12
scikit-learn==0.18.1
scipy==0.19.0
simplegeneric==0.8.1
six==1.10.0
tensorflow==1.0.0
traitlets==4.3.2
wcwidth==0.1.7
